idk what I changed

+++ src/main/java/doctor4t/ratsmischief/common/entity/ai/FollowOwnerRatGoal.java
@@ -5,11 +5,7 @@ import net.minecraft.block.BlockState;
 import net.minecraft.block.LeavesBlock;
 import net.minecraft.entity.LivingEntity;
 import net.minecraft.entity.ai.goal.Goal;
-import net.minecraft.entity.ai.pathing.BirdNavigation;
-import net.minecraft.entity.ai.pathing.EntityNavigation;
-import net.minecraft.entity.ai.pathing.LandPathNodeMaker;
-import net.minecraft.entity.ai.pathing.MobNavigation;
-import net.minecraft.entity.ai.pathing.PathNodeType;
+import net.minecraft.entity.ai.pathing.*;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.WorldView;
 
@@ -115,7 +111,7 @@ public class FollowOwnerRatGoal extends Goal {
 		} else if (!this.canTeleportTo(new BlockPos(x, y, z))) {
 			return false;
 		} else {
-			this.rat.refreshPositionAndAngles((double) x + 0.5D, (double) y, (double) z + 0.5D, this.rat.getYaw(), this.rat.getPitch());
+			this.rat.refreshPositionAndAngles((double) x + 0.5D, y, (double) z + 0.5D, this.rat.getYaw(), this.rat.getPitch());
 			this.navigation.stop();
 			return true;
 		}
